refactor: extract shared secret-activation response capture into a test utility
JobSecretActivationInjectionTest's response-writer interception and
controllable SecretCache exist because a resolved secret is only
ever observable through the mocked CommandResponseWriter — never in
the persisted event, since resolved values are deliberately never
logged, exported, or persisted. The upcoming cluster-variable
activation test needs the exact same scaffolding, and duplicating it
verbatim into a second file was flagged in review.

Extract it into SecretActivationResponseCapture (implements
SecretCache, exposes install/putSecret/failResolution/
awaitActivationResponse) and refactor JobSecretActivationInjectionTest
to use it. Pure refactor: field-for-field equivalent, same 10 tests
passing, no behavior change.
